# Optimizations for `write_cfg_data`