@PetarMax 2024-08-05
Before I forget, I think that it would be a good idea to have the set of pending nodes made explicit and maintained as they are modified - this should introduce a major speed-up since `get_steps()` looks for pending nodes, which means a traversal and analysis of the entire KCFG.