@PetarMax 2024-08-05
This looks like a good improvement, but I think that we might need to assess on engagement code. I am testing it with intermittent writing to disk (every `max-ftontier-parallel` iterations), and the speed-up is (expectedly) greater than the one shown here. Perhaps we could do both?