@nwatson22 2024-08-05
@PetarMax Thanks, that is a good suggestion for how we can solve the performance issues with the `APRProof` implementation of `get_steps()`.

Between the intermittent writing to disk and this change I suspect that the two would eat into each other's performance gain, but it sounds from what you're saying they stack at least somewhat. I will try to test this and #4563 together as well.